pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "waveid"
version = "0.1.0"
description = "Zero-Trust Digital Biometrics authentication for AI Agents"
requires-python = ">=3.10"
license = "Apache-2.0"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[tool.setuptools.packages.find]
where = ["src"]
